Create README.md
My ability is not only to craft efficient code but also oversee product lifecycles from conception to delivery while managing comprehensive program strategies. This amalgamation of technical proficiency and strategic oversight has significantly contributed to the overall success and growth of the company. A proficient Data and Business Analyst with a Master’s in Computer Applications and Master’s in Business Administration specializing in Python, Tableau, Power BI, Matplotlib, MySQL, and Microsoft Excel.